server/requirements.txt
# 核心框架
sanic==23.12.1
sanic-cors==2.2.0

# 数据库 ORM（支持 SQLite）
tortoise-orm==0.20.0
aiosqlite>=0.16.0,<0.18.0

# 数据验证和认证
pydantic==2.5.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 环境变量管理
python-dotenv==1.0.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 